+++ components/admin/ListUserComponent.jsx
@@ -6,13 +6,29 @@ import { useState, useEffect } from 'react';
 import { getHeaders } from '../../utils/konstanta';
 import Link from 'next/link';
 import { deleteData } from '../../utils/fetchApi';
+import Swal from 'sweetalert2';
+import axios from 'axios';
 
 const ListUserComponent = ({ dataUser, setUser }) => {
   const clickDelete = async (e, email) => {
     e.preventDefault();
-    // setUser([]);
-    const result = await deleteData('user', email);
-    alert('success', 'success', result.data.message);
+    const headers = getHeaders();
+    Swal.fire({
+      title: 'Apakah anda yakin untuk menghapus forum ini?',
+      icon: 'warning',
+      showDenyButton: true,
+      confirmButtonText: 'Ya',
+      denyButtonText: `Tidak`,
+    }).then(async (result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isConfirmed) {
+        Swal.fire('Data berhasil dihapus', '', 'success');
+        const result = await deleteData('user', email);
+        const updateData = await axios.get('https://familypreparation.up.railway.app/user', headers).then((res) => res.data.data);
+        console.log(updateData);
+        setUser(updateData);
+      }
+    });
   };
   return (
     <Card>
